# server/openapi.yaml
openapi: 3.0.0
info:
  version: "1"
  title: chefswap
  description: The API for chefswap
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/MATTHEWS13542/chefswap/1
  - url: http://localhost:3001
    description: development
paths:
...
  /api/v1/swaps/{accountUid}:
    parameters:
      - $ref: "#/components/parameters/AccountUid"
    get:
      summary: Get all swaps an account is a part of. Must be signed in with specified account.
      parameters:
        - in: query
          name: status
          description: The status of a swap (pending, ongoing, ended).
          schema:
            $ref: "#/components/schemas/SwapStatus"
        - in: query
          name: orderBy
          description: Specifies the first field to be used for ordering results. Second field (or only field if this param is not specified) is always accountUid.
          schema:
            type: string
            enum:
              - timeAsc
              - timeDesc
        - in: query
          name: key
          description: All swaps in response will be occur after a swap that has the specified properties according to the ordering set by orderBy. If ordering by time, key must include field "time". In all cases, field accountUid may optionally be provided.
          explode: true
          style: deepObject
          schema:
            oneOf:
              - type: object #no ordering (just order by uuidAsc)
                additionalProperties: false
                properties:
                  accountUid:
                    type: string
                    format: uuid
                    default: 00000000-0000-0000-0000-000000000000
              - type: object #ordering by time
                additionalProperties: false
                properties:
                  accountUid:
                    type: string
                    format: uuid
                    default: 00000000-0000-0000-0000-000000000000
                  time:
                    type: integer
                    format: int64
      responses:
        200:
          description: The swaps of the account corresponding to the provided uid matching status set in optional query param status.
          content:
            application/json:
              schema:
                type: object
                required:
                  - swaps
                properties:
                  swaps:
                    type: array
                    items:
                      $ref: "#/components/schemas/Swap"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: Create new swap. Must be signed in with specified account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Swap"
      responses:
        201:
          description: The newly created swap.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Swap"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
  /api/v1/swaps/{accountUid}/{swapperUid}/{requestTimestamp}:
    parameters:
      - $ref: "#/components/parameters/AccountUid"
      - name: swapperUid
        description: The uuid of an the swapper's account.
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: requestTimestamp
        description: Url encoded timestamp of corresponding to time the swap was requested.
        in: path
        required: true
        schema:
          type: string
          format: date-time
    get:
      summary: Get the specific swap between the specified accounts with specified requestTimestamp. Must be signed in with specified account.
      responses:
        200:
          description: The swap betweeen specified accounts with specified requestTimestamp.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Swap"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Delete the specific swap between the specified accounts with specified requestTimestamp. Must be signed in with specified account.
      responses:
        204:
          description: Successfully deleted swap.
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
  /api/v1/swaps/{accountUid}/{swapperUid}/{requestTimestamp}/{status}:
    parameters:
      - $ref: "#/components/parameters/AccountUid"
      - name: swapperUid
        description: The uuid of an the swapper's account
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: requestTimestamp
        description: Url encoded timestamp of corresponding to time the swap was requested.
        in: path
        required: true
        schema:
          type: string
          format: date-time
      - name: status
        description: The status of a swap (pending, ongoing, ended).
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/SwapStatus"
    put:
      summary: Change the status of a swap. Must be signed in with one of the specified accounts and status must correspond with future status.
      description: Can be set to ongoing if currently pending signed in as account that did no request. Can be set to ended currently ongoing and signed in with either account.
      responses:
        200:
          description: The new status of the swap.
          content:
            application/json:
              schema:
                type: object
                required:
                  - swapStatus
                properties:
                  swapStatus:
                    $ref: "#/components/schemas/SwapStatus"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
...
    SwapStatus:
      type: string
      enum:
        - pending
        - ongoing
        - ended
    Swap:
      type: object
      required:
        - requesterUid
        - requesteeUid
        - requestTimestamp
        - status
      properties:
        requesterUid:
          readOnly: true
          type: string
          format: uuid
        requesteeUid:
          type: string
          format: uuid
        requestTimestamp:
          readOnly: true
          type: string
          format: date-time
        acceptTimestamp:
          readOnly: true
          type: string
          format: date-time
        endTimestamp:
          readOnly: true
          type: string
          format: date-time
        status:
          readOnly: true
          allOf:
            - $ref: "#/components/schemas/SwapStatus"
    Errors:
      type: array
      items:
        $ref: "#/components/schemas/Error"
    Error:
      type: object
      required:
        - path
        - message
        - detail
      properties:
        path:
          type: string
        message:
          type: string
        detail:
          type: string
...
  responses:
    BadRequest:
      description: request format is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Errors"
    Unauthorized:
      description: action requires authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Errors"
    Forbidden:
      description: insufficient permission for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Errors"
    NotFound:
      description: specified resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Errors"
...
    InternalServerError:
      description: error processing request on server
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Errors"
  parameters:
    AccountUid:
      name: accountUid
      description: The uuid of an account
      in: path
      required: true
      schema:
        type: string
        format: uuid